Supports ArgCD in a separate cluster from the workload (#1)
TL;DR
-----

Enables deploying to a cluster other than the Argo CD cluster

Details
-------

Reorganizes the manifests so that Argo CD can run on a different
cluster from the Postfacto application. Right now I "borrow" the
credentials for the target cluster from Argo CD to run preflight
checks. I did this to simplify things and since Argo would have
the credentials we need on the workload cluster.

In this model, preflights run on the ArgoCD cluster and Argo
will deploy two different applications are if the preflights are 
successful--one to set up the secrets the chart install depends
on, and another to deploy the chart. I run the preflights on
the Argo cluster so that nothing is change on the destination
cluster if the pre-flights fail.