Documents for sharing with others (#2)
TL;DR
-----

Explains how to use the repository and how it's licensed

Details
-------

Adds a license and README so that others can use the repo and
I'll remember how to use it in the future. The demo is licensed
under Apache 2.0 so you can use it as a baseline for your own
work, subject to the constraints of that license.